Modified the Components Architecture
Added WebClient as a root component, it includes NavBar and
ActionManager. The ActionManager serves as a place for components
switching.
With regard to the login process, for now it's a fake one, we just allow
any user who filled the fields and submit, to login. After the login,
Course will be the component to be displayed.