Minimize The Initial Data Come In the Base Module
Try to make the initial data that would be stored in the database as
small as possible. To do that romve 'data/res_currency_data.xml',
'data/res_country_data.xml', 'data/res.country.state.csv', and
'data/ir_module_module.xml' from odoo/addons/base/__manifest__.py.

Still there might be more that can be extracted out, but for now that is
enough.

On this commit there is also an addition of heroku database
initialization file (db_initializer.py). Actually this file is not
specific for heroku platform, but it is useless in the normal
circumstances; because normally database initialization will occur automatically.