Remove The Stratbootstrap Dependencies
Stratbootstrap is the one who creates boostrap themes. Slice03 theme
happend to be created by them. The theme has dependency on bootstrap
library, and it is usually bundled together. And since odoo itself has a
bootstrap library bundled in the web module, we decided to delete the
one that come with the theme.